+++ src/main/java/net/grid/vampiresdelight/common/registry/VDBlocks.java
@@ -122,41 +122,41 @@ public class VDBlocks {
     public static final RegistryObject<Block> ORCHID_CAKE = BLOCKS.register("orchid_cake",
             () -> new ConsumableCakeBlock(Block.Properties.copy(Blocks.CAKE).mapColor(MapColor.TERRACOTTA_PURPLE), VDItems.BLOOD_PIE_SLICE));
     public static final RegistryObject<Block> ORCHID_CANDLE_CAKE = BLOCKS.register("orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> WHITE_ORCHID_CANDLE_CAKE = BLOCKS.register("white_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.WHITE_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.WHITE_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> ORANGE_ORCHID_CANDLE_CAKE = BLOCKS.register("orange_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.ORANGE_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.ORANGE_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> MAGENTA_ORCHID_CANDLE_CAKE = BLOCKS.register("magenta_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.MAGENTA_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.MAGENTA_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> LIGHT_BLUE_ORCHID_CANDLE_CAKE = BLOCKS.register("light_blue_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.LIGHT_BLUE_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.LIGHT_BLUE_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> YELLOW_ORCHID_CANDLE_CAKE = BLOCKS.register("yellow_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.YELLOW_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.YELLOW_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> LIME_ORCHID_CANDLE_CAKE = BLOCKS.register("lime_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.LIME_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.LIME_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> PINK_ORCHID_CANDLE_CAKE = BLOCKS.register("pink_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.PINK_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.PINK_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> GRAY_ORCHID_CANDLE_CAKE = BLOCKS.register("gray_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.GRAY_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.GRAY_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> LIGHT_GRAY_ORCHID_CANDLE_CAKE = BLOCKS.register("light_gray_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.LIGHT_GRAY_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.LIGHT_GRAY_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> CYAN_ORCHID_CANDLE_CAKE = BLOCKS.register("cyan_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.CYAN_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.CYAN_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> PURPLE_ORCHID_CANDLE_CAKE = BLOCKS.register("purple_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.PURPLE_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.PURPLE_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> BLUE_ORCHID_CANDLE_CAKE = BLOCKS.register("blue_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.BLUE_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.BLUE_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> BROWN_ORCHID_CANDLE_CAKE = BLOCKS.register("brown_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.BROWN_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.BROWN_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> GREEN_ORCHID_CANDLE_CAKE = BLOCKS.register("green_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.GREEN_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.GREEN_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> RED_ORCHID_CANDLE_CAKE = BLOCKS.register("red_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.RED_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.RED_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
     public static final RegistryObject<Block> BLACK_ORCHID_CANDLE_CAKE = BLOCKS.register("black_orchid_candle_cake",
-            () -> candleCakeBlock(Blocks.BLACK_CANDLE, ORCHID_CAKE, MapColor.TERRACOTTA_PURPLE));
+            () -> candleCakeBlock(Blocks.BLACK_CANDLE, (ConsumableCakeBlock) ORCHID_CAKE.get(), MapColor.TERRACOTTA_PURPLE));
 
-    public static ConsumableCandleCakeBlock candleCakeBlock(Block candleBlock, Supplier<Block> cakeBlock, MapColor mapColor) {
+    public static ConsumableCandleCakeBlock candleCakeBlock(Block candleBlock, ConsumableCakeBlock cakeBlock, MapColor mapColor) {
         return new ConsumableCandleCakeBlock(Block.Properties.copy(Blocks.CAKE)
                 .mapColor(mapColor).lightLevel(litBlockEmission(3)), cakeBlock, candleBlock);
     }